refactor(websocket): extract DirectInferenceHandler from the runner
T5 of docs/websocket-runner-refactor-plan.md. The one-shot model calls —
the streamed `inference` command and the direct text / media / speech
generation the sketch, timeline and Studio surfaces drive over RPC — were
reachable from all 8,800 lines of the runner alongside job execution, chat
and the command switch. They depend on almost nothing the connection owns,
so they move behind `ClientSession` with the two things they genuinely need
declared in the constructor.

The second half is the `registerAbort` seam. `rpcAborts` stays on the host,
because `stop` and disconnect must reach every in-flight provider call, but
nothing outside the host touches the set any more: a caller registers a
controller and gets its deregistration back. T6 takes the same dependency
when the command switch moves, so the set keeps one owner.

Same frames in, same messages out, same abort semantics.
